[#90] LikeController의 메서드 MemesController로 이관
