+++ backend/memetory/src/main/java/com/example/memetory/domain/memes/controller/MemesApi.java
@@ -115,4 +115,38 @@ public interface MemesApi {
 		)
 	})
 	ResponseEntity<ResultResponse> findAllMemes(Pageable pageable);
+
+	@Operation(
+		summary = "meme`s 좋아요 등록",
+		description = "공유된 meme`s 게시물에 좋아요를 등록한다.",
+		security = {@SecurityRequirement(name = "access_token")}
+	)
+	@ApiResponses(value = {
+		@ApiResponse(
+			responseCode = "201",
+			description = "좋아요 등록!"
+		)
+	})
+	ResponseEntity<ResultResponse> registerLike(
+		@Parameter(hidden = true) String email,
+		@Parameter(in = ParameterIn.PATH, description = "밈스 아이디", required = true)
+		Long memesId
+	);
+
+	@Operation(
+		summary = "meme`s 좋아요 취소",
+		description = "공유된 meme`s 게시물에 등록된 좋아요를 취소한다.",
+		security = {@SecurityRequirement(name = "access_token")}
+	)
+	@ApiResponses(value = {
+		@ApiResponse(
+			responseCode = "200",
+			description = "좋아요 취소!"
+		)
+	})
+	ResponseEntity<ResultResponse> cancelLike(
+		@Parameter(hidden = true) String email,
+		@Parameter(in = ParameterIn.PATH, description = "밈스 아이디", required = true)
+		Long memesId
+	);
 }
